+++ ShaderLibrary/Surface.hlsl
@@ -1,25 +1,25 @@
 #ifndef _GRP_SURFACE
 #define _GRP_SURFACE
 
-half _GlobalShadowStrength;
+float _GlobalShadowStrength;
 
 struct Surface
 {
 	float3 position;
-	half3 normal;
-	half3 interpolatedNormal;
-	half3 viewDirection;
-	half depth;
-	half3 color;
-	half alpha;
-	half metallic;
-	half smoothness;
-	half fresnel;
-	half dither;
-    half shadowStrength;
+	float3 normal;
+	float3 interpolatedNormal;
+	float3 viewDirection;
+	float depth;
+	float3 color;
+	float alpha;
+	float metallic;
+	float smoothness;
+	float fresnel;
+	float dither;
+    float shadowStrength;
 };
 
-Surface MakeSurface(half4 diffuse, float3 positionWS, half3 interpolatedNormalWS, half metallic, half smoothness, half fresnel)
+Surface MakeSurface(float4 diffuse, float3 positionWS, float3 interpolatedNormalWS, float metallic, float smoothness, float fresnel)
 {
 	Surface surface;
 	surface.position = positionWS;
@@ -37,7 +37,7 @@ Surface MakeSurface(half4 diffuse, float3 positionWS, half3 interpolatedNormalWS
 	return surface;
 }
 
-Surface MakeSurface(half4 diffuse, float3 positionWS, half3 interpolatedNormalWS, half metallic, half smoothness, half fresnel, float2 ditherUV)
+Surface MakeSurface(float4 diffuse, float3 positionWS, float3 interpolatedNormalWS, float metallic, float smoothness, float fresnel, float2 ditherUV)
 {
 	Surface surface = MakeSurface(diffuse, positionWS, interpolatedNormalWS, metallic, smoothness, fresnel);
 	surface.dither = InterleavedGradientNoise(ditherUV, 0);
